Fix: workflow placeholder issue
